test good

+++ data_structure.h
@@ -38,7 +38,7 @@ void create_new_element(void* s, size_t len, size_t position, vector_dict* vec, 
 
 vector_dict* Create(void* v_str, type_info* type);
 
-size_t find(void* v_find_string, vector_dict* vec, int n, type_info* type);
+int find(void* v_find_string, vector_dict* vec, int n, type_info* type);
 
 void free_data_struct(vector_dict* vec);
